fix member reg_mch_id

+++ core/domain/interface/member/invitation_manager.go
@@ -27,7 +27,7 @@ type IInvitationManager interface {
 	GetInvitationMeMember() *Member
 
 	// 更换邀请人,async是否异步更新
-	UpdateInviter(inviterId int64, sync bool) error
+	UpdateInviter(inviterId int, sync bool) error
 
 	// 屏蔽
 	Shield(memberId int) error
